fix CarController, implement SRP
