+++ src/main/java/id/ac/ui/cs/advprog/eshop/controller/ProductController.java
@@ -58,52 +58,4 @@ public class ProductController {
         service.delete(productToDelete);
         return "redirect:../list";
     }
-
-    @Controller
-    @RequestMapping("/car")
-    class CarController extends ProductController{
-        @Autowired
-        private CarServiceImpl carService;
-
-        @GetMapping("/createCar")
-        public String createCarPage(Model model){
-            Car car = new Car();
-            model.addAttribute("car", car);
-            return "createCar";
-        }
-
-        @PostMapping("/createCar")
-        public String createCarPost(@ModelAttribute Car car, Model model){
-            carService.create(car);
-            return "redirect:listCar";
-        }
-
-        @GetMapping("/listCar")
-        public String carListPage(Model model){
-            List<Car> allCars = carService.findAll();
-            model.addAttribute("cars", allCars);
-            return "carList";
-        }
-
-        @GetMapping("/editCar/{carId}")
-        public String editCarPage(@PathVariable String carId, Model model){
-            Car car = carService.findById(carId);
-            model.addAttribute("car", car);
-            return "editCar";
-        }
-
-        @PostMapping("/editCar")
-        public String editCarPost(@ModelAttribute Car car, Model model){
-            System.out.println(car.getCarId());
-            carService.update(car.getCarId(), car);
-
-            return "redirect:listCar";
-        }
-
-        @PostMapping("/deleteCar")
-        public String deleteCar(@RequestParam("carId") String carId){
-            carService.deleteCarById(carId);
-            return "redirect:listCar";
-        }
-    }
 }
